requirements.txt
PyYAML>=6.0.1
requests>=2.31.0
feedparser>=6.0.11
